# sdk/openapi.yaml
openapi: 3.1.0
info:
  title: Ziro OS API
  version: "1"
  summary: Manage a Ziro OS host or cluster master (ziroctl api)
  description: |
    The API server (`ziroctl api start`) listens on loopback by default (HTTPS optional).
    Authenticate with `Authorization: Bearer <token>` (`ziroctl api token`, or scoped tokens from
    `ziroctl api token create <name> --role viewer|operator|admin`).

    Roles: each operation states the role it needs in `x-ziro-role` (the server enforces exactly
    that). **viewer** reads; **operator** also runs day-to-day changes (services, containers, DNS
    records, scaling, node cordon/drain); **admin** may do everything, including what grants or
    removes access, installs software, deletes data or changes where traffic goes. Requests are
    rate limited (120/min per client, 600/min per token) and every change is audited
    (`GET /api/v1/events` streams the audit trail). Errors are `{"status": "error", "code", "message"}`.

    Long jobs (module enable, app deploy) answer `202 Accepted` and run in the background; poll the
    matching GET. The Go client is `github.com/ziro-os/ziro-os/sdk/client`.
  license:
    name: MIT
servers:
  - url: https://127.0.0.1:8443
security:
  - bearer: []
...
paths:
...
  /api/v1/gateway/routes:
    get:
      x-ziro-role: viewer
      tags: [gateway]
      summary: Routes with their resolved upstreams
      responses:
        "200":
          description: Routes
          content:
            application/json:
              schema: { type: array, items: { $ref: "#/components/schemas/GatewayRouteState" } }
    post:
      x-ziro-role: admin
      tags: [gateway]
      summary: Create or replace a route (admin; live within a second)
      requestBody: { $ref: "#/components/requestBodies/GatewayRoute" }
      responses:
        "200": { description: The stored route, content: { application/json: { schema: { $ref: "#/components/schemas/GatewayRoute" } } } }
        "400": { $ref: "#/components/responses/Message" }
  /api/v1/gateway/routes/{name}:
    get:
      x-ziro-role: viewer
      tags: [gateway]
      summary: One route
      parameters: [{ $ref: "#/components/parameters/Name" }]
      responses:
        "200": { description: Route, content: { application/json: { schema: { $ref: "#/components/schemas/GatewayRouteState" } } } }
    put:
      x-ziro-role: admin
      tags: [gateway]
      summary: Create or replace this route (admin)
      parameters: [{ $ref: "#/components/parameters/Name" }]
      requestBody: { $ref: "#/components/requestBodies/GatewayRoute" }
      responses:
        "200": { description: The stored route, content: { application/json: { schema: { $ref: "#/components/schemas/GatewayRoute" } } } }
        "400": { $ref: "#/components/responses/Message" }
    delete:
      x-ziro-role: admin
      tags: [gateway]
      summary: Remove this route (admin)
      parameters: [{ $ref: "#/components/parameters/Name" }]
      responses:
        "200": { $ref: "#/components/responses/Message" }
...
components:
  securitySchemes:
    bearer:
      type: http
      scheme: bearer
  parameters:
...
    Name:
      name: name
      in: path
      required: true
      schema: { type: string, pattern: "^[a-z0-9][a-z0-9_.-]{0,62}$" }
  requestBodies:
...
    GatewayRoute:
      required: true
      content:
        application/json:
          schema: { $ref: "#/components/schemas/GatewayRoute" }
  responses:
...
    Message:
      description: Result
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Message" }
  schemas:
...
    Message:
      type: object
      properties:
        status: { type: string, enum: [ok, error, accepted] }
        code: { type: string, description: "errors: invalid, unauthorized, forbidden, not_found, method_not_allowed, conflict, too_large, rate_limited, internal" }
        message: { type: string }
...
    GatewayUpstream:
      type: object
      properties:
        app: { type: string }
        address: { type: string, description: "IP:port" }
        weight: { type: integer, minimum: 1, maximum: 1000 }
    GatewayRoute:
      type: object
      description: A routing rule. Exactly one of to, redirect, respond.
      properties:
        name: { type: string }
        kind: { type: string, enum: [http, tcp, tls], default: http }
        hosts: { type: array, items: { type: string }, description: 'exact names or "*.domain" (one label)' }
        path_prefix: { type: string, default: / }
        path_exact: { type: boolean }
        methods: { type: array, items: { type: string } }
        headers: { type: object, additionalProperties: { type: string } }
        listen: { type: integer, description: port of a tcp route }
        to: { type: array, items: { $ref: "#/components/schemas/GatewayUpstream" } }
        lb: { type: string, enum: [round_robin, least_conn, ip_hash, cookie] }
        health:
          type: object
          properties:
            path: { type: string }
            interval: { type: string, examples: [10s] }
            expect: { type: integer }
        redirect: { type: string, description: '"{uri}" is the request URI' }
        redirect_code: { type: integer, enum: [301, 302, 307, 308] }
        respond:
          type: object
          properties:
            status: { type: integer }
            body: { type: string }
        strip_prefix: { type: string }
        request_headers: { $ref: "#/components/schemas/HeaderRules" }
        response_headers: { $ref: "#/components/schemas/HeaderRules" }
        retries: { type: integer, minimum: 0, maximum: 5 }
        timeout: { type: string }
        compress: { type: boolean }
        tls: { type: string, examples: [auto, off, internal, "cert:name"] }
        allow_cidrs: { type: array, items: { type: string } }
        rate_rps: { type: integer }
        max_body_mb: { type: integer }
        basic_auth: { type: object, additionalProperties: { type: string }, description: user -> bcrypt hash }
    HeaderRules:
      type: object
      properties:
        set: { type: object, additionalProperties: { type: string } }
        remove: { type: array, items: { type: string } }
    GatewayRouteState:
      allOf:
        - $ref: "#/components/schemas/GatewayRoute"
        - type: object
          properties:
            upstreams: { type: array, items: { type: string } }
            targets:
              type: array
              items:
                type: object
                properties:
                  addr: { type: string }
                  weight: { type: integer }
...